[backend] refactor(document):Extract document files from services into their own document module
