+++ apps/portal-api/src/modules/document/document.security.ts
@@ -1,13 +1,13 @@
 import {
   ServiceDefinitionIdentifier,
   ServiceRestriction,
-} from '../../../__generated__/resolvers-types';
-import { ServiceInstanceId } from '../../../model/kanel/public/ServiceInstance';
-import { UserLoadUserBy } from '../../../model/user';
-import { isUserGranted } from '../../../security/access';
-import { ErrorCode } from '../../../utils/error/error.code';
-import { loadCapabilities } from '../../user_service/user-service-capability/user-service-capability.helper';
-import { loadServiceDefinitionByServiceInstance } from '../service-instance.domain';
+} from '../../__generated__/resolvers-types';
+import { ServiceInstanceId } from '../../model/kanel/public/ServiceInstance';
+import { UserLoadUserBy } from '../../model/user';
+import { isUserGranted } from '../../security/access';
+import { ErrorCode } from '../../utils/error/error.code';
+import { loadServiceDefinitionByServiceInstance } from '../services/service-instance.domain';
+import { loadCapabilities } from '../user_service/user-service-capability/user-service-capability.helper';
 
 export const isUserRestrictedToActiveDocument = async (
   user: UserLoadUserBy,
