+++ apps/portal-api/src/modules/document/openaev/scenarios/scenarios.resolver.ts
@@ -1,13 +1,13 @@
-import { Resolvers } from '../../../../../__generated__/resolvers-types';
-import { ServiceInstanceId } from '../../../../../model/kanel/public/ServiceInstance';
-import { subscriptionApp } from '../../../../subcription/subscription.app';
-import { useCaseDomain } from '../../../../use-case/use-case.domain';
-import { getServiceInstance } from '../../../service-instance.domain';
+import { Resolvers } from '../../../../__generated__/resolvers-types';
+import { ServiceInstanceId } from '../../../../model/kanel/public/ServiceInstance';
+import { getServiceInstance } from '../../../services/service-instance.domain';
+import { subscriptionApp } from '../../../subcription/subscription.app';
+import { useCaseDomain } from '../../../use-case/use-case.domain';
 import { DocumentChildrenDomain } from '../../domain/document.children.domain';
 import { DocumentDomain } from '../../domain/document.domain';
 
 const resolvers: Resolvers = {
-  CustomDashboard: {
+  OpenAEVScenario: {
     use_cases: ({ id }) => useCaseDomain.loadUseCasesByDocumentId(id),
     children_documents: ({ id }) =>
       DocumentChildrenDomain.loadImagesByDocumentId(id),
@@ -16,7 +16,7 @@ const resolvers: Resolvers = {
       DocumentDomain.loadUploaderOrganization(id),
     service_instance: ({ service_instance_id }, _) =>
       getServiceInstance(service_instance_id as ServiceInstanceId),
-    subscription: async ({ service_instance_id }, _, context) =>
+    subscription: ({ service_instance_id }, _, context) =>
       subscriptionApp.loadSubscriptionModel(context.user, service_instance_id),
   },
 };
