+++ apps/portal-api/src/modules/document/opencti/custom-dashboards/custom-dashboards.graphql
