+++ apps/portal-api/src/modules/document/opencti/integrations/integrations.graphql
