+++ apps/portal-api/src/modules/document/visualize-document-endpoint.ts
@@ -1,16 +1,16 @@
 import cors from 'cors';
 import { fromGlobalId } from 'graphql-relay/node/node.js';
 import { Readable } from 'stream';
-import { requestContext } from '../../../context/request.context';
-import { DocumentId } from '../../../model/kanel/public/Document';
-import ServiceDefinition from '../../../model/kanel/public/ServiceDefinition';
-import { ServiceInstanceId } from '../../../model/kanel/public/ServiceInstance';
-import { PortalContext } from '../../../model/portal-context';
-import { MinIOClient } from '../../../thirdparty/minio/client';
-import { logApp } from '../../../utils/app-logger.util';
-import { NotFoundError } from '../../../utils/error/error.util';
-import { extractId } from '../../../utils/utils';
-import { loadServiceDefinitionByServiceInstance } from '../service-instance.domain';
+import { requestContext } from '../../context/request.context';
+import { DocumentId } from '../../model/kanel/public/Document';
+import ServiceDefinition from '../../model/kanel/public/ServiceDefinition';
+import { ServiceInstanceId } from '../../model/kanel/public/ServiceInstance';
+import { PortalContext } from '../../model/portal-context';
+import { MinIOClient } from '../../thirdparty/minio/client';
+import { logApp } from '../../utils/app-logger.util';
+import { NotFoundError } from '../../utils/error/error.util';
+import { extractId } from '../../utils/utils';
+import { loadServiceDefinitionByServiceInstance } from '../services/service-instance.domain';
 import { DocumentDomain } from './domain/document.domain';
 
 export const documentVisualizeEndpoint = (app) => {
