+++ apps/portal-api/src/modules/ingest-manifest/ingest-manifest.model.ts
@@ -1,4 +1,4 @@
-import { Connector } from '../services/document/opencti/integrations/integrations.model';
+import { Connector } from '../document/opencti/integrations/integrations.model';
 
 export interface ManifestInformation extends Partial<Connector> {
   use_cases: string[];
