+++ apps/portal-api/src/modules/telemetry/telemetry.app.test.ts
@@ -29,10 +29,10 @@ import {
   ServiceConfigurationStatus,
 } from '../../__generated__/resolvers-types';
 import type { ServiceInstanceId } from '../../model/kanel/public/ServiceInstance';
-import { DocumentApp } from '../services/document/document.app';
-import { deleteDocuments } from '../services/document/document.helper';
-import * as DocumentUploadsHelper from '../services/document/document.uploads.helper';
-import { INTEGRATION_SERVICE_INSTANCE_ID } from '../services/document/opencti/integrations/integrations.model';
+import { DocumentApp } from '../document/document.app';
+import { deleteDocuments } from '../document/document.helper';
+import * as DocumentUploadsHelper from '../document/document.uploads.helper';
+import { INTEGRATION_SERVICE_INSTANCE_ID } from '../document/opencti/integrations/integrations.model';
 import * as serviceInstanceDomain from '../services/service-instance.domain';
 
 // Mock the ES Client
