+++ apps/portal-api/src/modules/telemetry/telemetry.helper.ts
@@ -11,7 +11,7 @@ import { UserId } from '../../model/kanel/public/User';
 import { loadOrganizationBy } from '../organizations/organizations.domain';
 
 import { requestContext } from '../../context/request.context';
-import { DocumentMetadataDomain } from '../services/document/domain/document.metadata.domain';
+import { DocumentMetadataDomain } from '../document/domain/document.metadata.domain';
 import { loadServiceDefinitionByServiceInstance } from '../services/service-instance.domain';
 import {
   TelemetryEventService,
