+++ apps/portal-api/src/modules/xtm-suite-roadmap/epic.app.test.ts
@@ -14,9 +14,9 @@ import Document, { DocumentId } from '../../model/kanel/public/Document';
 import Epic, { EpicId } from '../../model/kanel/public/Epic';
 import { OrganizationId } from '../../model/kanel/public/Organization';
 import { MinIOClient } from '../../thirdparty/minio/client';
-import { DocumentApp } from '../services/document/document.app';
-import * as DocumentUploadsHelper from '../services/document/document.uploads.helper';
-import { DocumentDomain } from '../services/document/domain/document.domain';
+import { DocumentApp } from '../document/document.app';
+import * as DocumentUploadsHelper from '../document/document.uploads.helper';
+import { DocumentDomain } from '../document/domain/document.domain';
 import * as ServiceInstanceDomain from '../services/service-instance.domain';
 import { EpicApp } from './epic.app';
 import { EpicDomain } from './epic.domain';
