+++ apps/portal-api/src/modules/xtm-suite-roadmap/epic.resolver.ts
@@ -4,7 +4,7 @@ import { DocumentId } from '../../model/kanel/public/Document';
 import Epic, { EpicId } from '../../model/kanel/public/Epic';
 import { UnknownErrorCode } from '../../utils/error/error.code';
 import { mapToGraphQLError } from '../../utils/error/error.mapping';
-import { DocumentDomain } from '../services/document/domain/document.domain';
+import { DocumentDomain } from '../document/domain/document.domain';
 import { EpicApp } from './epic.app';
 
 const resolvers: Resolvers = {
